python/concurrency/requirements.txt
numpy==1.26.0
requests==2.31.0
Pillow==10.0.1
